# Speedup `GDScript::get_script_property_list` by not using an intermediate Vector